requirements.txt
arrow==1.3.0
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
isort==5.13.2
Jinja2==3.1.4
jinja2-time==0.2.0
make==0.1.6.post2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
types-python-dateutil==2.9.0.20240906
uritemplate==4.1.1
